Fold #228/#229 test coverage into PyCseMachinePlugin.test.ts
The PR description claimed PyCseMachinePlugin.test.ts had a pre-existing
compile error (globalNames missing on CseSerializedEnvFrame) that
justified splitting new coverage into separate files. That claim was
wrong: it was caused by a stale local node_modules install
(@sourceacademy/common-cse-machine@0.1.0 instead of the yarn.lock-pinned
0.2.0, which added globalNames). On a correct install the file compiles
and passes cleanly, so there's no reason to deviate from this repo's
one-file-per-module test convention.

Moves the synthetic-node and builtin-display regression tests into
PyCseMachinePlugin.test.ts's existing describe blocks and removes the
two standalone files.

Thanks @Shrey for catching this.
